image-search-api/pyproject.toml
[project]
name = "image-search-api"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "fastapi (>=0.126.0,<0.127.0)",
    "pillow (>=12.0.0,<13.0.0)",
    "boto3 (>=1.42.14,<2.0.0)",
    "dependency-injector (>=4.48.3,<5.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)",
    "python-multipart (>=0.0.21,<0.0.22)",
    "numpy (>=1.26.0,<3.0.0)",
    "opensearch-py (>=2.5.0,<3.0.0)",
    "tritonclient (==2.59)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
